Moved all of the score period modals into one file to condense things since they're really not that big
